Bugfix for Cpermadeath re: Legend Sword
In vanilla FE, most people don't actually *use* the Legend Sword as
a weapon, but you can if you want. On Cpermadeath/deader, since your
character is wiped out and you lose your equipment, it's important
to try to retain the Legend Sword, since you need it to be able to
forge at Kokkol's (e.g. for the objective). So, code in permadeath.f4c
attempts to salvage the Legend Sword if your character permadies in
battle. This code... doesn't always work. The goal here is to help
the code always work.

The main thing that's wonky is that the code in permadeath.f4c runs
a different number of times depending on how you end the battle.
If you run away, or if you win the battle but the battle does not
allow "spoils" (meaning experience, gp, etc.), then this code only
runs once. If you perish, the code doesn't run at all (including if
it's a fake-perish, e.g. Dark Elf, unless it's Cpermadeader, where it
runs once). Otherwise, it runs *twice*. For everything but the Legend
Sword, it's not a big deal (though it does a non-trivial amount of
useless stuff), but the Legend Sword code is intended to work around
it running twice, and fails for two reasons.

First, the game re-copies the battle inventory to your field inventory
right before the permadeath code runs. The Legend Sword code tries to
stash the Legend Sword in your *field* inventory... so it gets
immediately overwitten by the second copying.

Second, in the event that you bag-full during the above, then the code
tries to put it in the spoils, which is fine... unless the game then
gives you drops or a MIAB reward. The former happens in-between the
two runs of the permadeath code, so a drop overwrites the sword.
The latter happens after the sword is potentially put into the spoils,
and... then the MIAB reward overwrites it. The permadeath code actually
has this as failure condition: if it's a MIAB, you're out of luck, so
just hope it works out. There's a problem in this code, though:
it checks "$c9", but the direct register is $0000 here instead of
$0600, so it checks what Aexoden's disassembly calls an unknown value.
Most of the time it just falls through to the normal spoils loop,
puts the sword there, and then the sword is deleted by the code in
rewards.f4c (Rewards__MIABContents). I think the $c9 thing happened
because earlier in the code, we temporarily have $0600 in the direct
register to possibly swap party leader (and it gets set back).

So, the bugfix goes as follows:
- at L32 of permadeath.f4c, swap the Legend Sword check and the "zero
out" of the current character's axtor ID, and then insert code that
checks, in order:
1. if you are running away (if yes, check for the Legend Sword),
2. if you finished a battle without exp/gp/etc. (if yes, also check
for the Legend Sword), and
3. if the current axtor ID is 0 (if yes, *now* we check for the
Legend Sword).
In the first two cases, this code only runs once, so we need to
possibly salvage the Legend Sword. Otherwise, you just finished a
battle that has spoils, so we need to *skip* the Legend Sword check
until the second time. The way we do this is by still zeroing out the
axtor ID regardless of the above checks and then checking if the axtor
ID is already 0 the second time (it can never *start* 0 unless the
slot is empty, in which case this code runs... and does nothing).

- right before the spoils loop starts (L95 of permadeath.f4c), we
change "lda $c6" to "lda $06c6", to correctly check whether or not
we're in a MIAB encounter.

- at the end of permadeath.f4c, we relabel %SpoilsUnavailable to
%MIABSpoils and then put the Legend Sword into the *second* spoils
slot ($1805) before returning.

- in rewards.f4c at L451 (in vanilla FE, L485 in the fork repo),
we check the second spoils slot in case specifically the Legend Sword
is there; if it is, we don't zero it out, so that the spoils menu
will give both the MIAB reward and the Legend Sword.

It's possible that this bugfix will need to be reworked for some
reason, so watch for future updates.