Adding Cbrieflydead and Cmostlydead
warlink suggested "any downed party members at the end of
battle get sent to the Tower of Wishes instead of being removed from
existence." Many of us agreed that this idea was good! Moreover,
we could have a different version where your character gets sent
back to the Tower but at base level, so I made that too.

The base of these flags is, of course, Cpermadeath. (See the bugfix
commit for the changes there, that have propagated to this change.)
Instead of just deleting the character (by zeroing out the axtor
ID in their field record), we do that and *also*...
- jump partway into CharEx__PushPartySlotToShadowParty (at the new
label CharEx__PushPartySlotToShadowPartyAxtorKnown) to push the
swooned character's record to the shadow party, when on Cbrieflydead,
or
- call CharEx__InitializeAxtor to put that character's base stats into
the right place in the shadow party, when on Cmostlydead.

If the swooned character has the Legend Sword, then we care if the
sword would disappear if the character resets or is unreachable. Since
Cbye works by preventing you from accessing the shadow party
characters in the Tower of Wishes, we need to check for the Legend
Sword on Cbrieflydead if Cbye is also on, and then obviously on
Cmostlydead. The Cpermadeath changes apply here too, to only check
for the Legend Sword when it'll actually work (see the other commit
for a long explanation).

The code for this stuff is in swoon_to_wishes.f4c; character_rando.py
has new code for adding the Legend check toggle; post_battle.f4c has
new hooks at the same place Cpermadeath hooks in; the new label
is in eventextensions_misc.f4c; uispec and flagspec and flag
descriptions have been updated as well.