Typo correction in commit 7ca0770
The end of the commit should read something like:

The code for this stuff is in swoon_to_wishes.f4c; character_rando.py
has new code for adding the Legend check toggle; post_battle.f4c has
new hooks at the same place Cpermadeath hooks in; the new label
is in character_expansion.f4c; uispec and flagspec and flag
descriptions have been updated as well. [The previous commit had
eventextensions_misc.f4c instead of character_expansion.f4c.]

A line (L580) in eventextensions_misc.f4c has been commented out,
because it seems to be a duplicate call to CharEx__PushShadowPartySlot
when that subroutine is called in CharEx_InitializeAxtor.
We'll see if that causes issues in the future!

Moral of this story: write better documentation outside of commit
messages and/or don't make mistakes.