Fixing -exp:moonbonus
Originally, -exp:moonbonus[100/200] worked by looking at the map plane
(if it's 2, then you're on the moon's surface, so you should get bonus
exp) and at the map ID (if the map plane is 1 and the map ID is at
least 0x5A, then you should be in a moon dungeon map). Map plane is
$1701, (low byte of the) map ID is $1702.

... The issue is that exiting to the world map does not change your
map ID, so e.g. exiting from the Hook route to the underworld means
your map plane is 1 and your map ID is bigger than 0x5A, so that
eggs (!) give bonus exp. The solution is to remember that the map
plane only gives the high byte of the map ID when the map *area*
($1700) is 3 (for dungeons). Moreover, we can simply check the map
area for 2 to test for the moon's surface, so the correct/simplified
check is:
1. Check the map area for 2 (on moon's surface); if yes, then give
bonus exp.
2. Check if it's 3 (in a dungeon); if yes, then check the two-byte
map ID ($1701-02) and if yes, then give bonus exp.

A couple of addresses were modified to remove the 7e bank byte,
to save an extra couple of bytes. Turns out it was also fewer bytes
to just put the two-byte map ID into a register and check that
instead of checking separately.