Adding -prng: flags (and updating main_info.md)
The random number generation in FF4 is pseudo-random: there is a table
of 256 bytes (0-255, shuffled) that get used one at a time. These
flags change what the PRNG table is.

-prng:shuffle just shuffles the pre-existing table.

-prng:random fully randomizes each byte of the table, with minimal
regard for duplicate values. To prevent battle softlocks, there's a
check to ensure all slots are targettable, since the game just tries
over and over until it finds a valid target. (This check is
implemented as a function.)

-prng:consecutive gives the values 0 to 255, in ascending order.

-prng:mostlysingle uses a single number for almost all of the values,
except 12 entries that are near the chosen value, again to ensure
that all slots in battle are targettable. Choosing a block of 13
numbers including the chosen number accomplishes this goal, due to
how the game gets battle rng. The different entries are spread out
through the table (and are in a random order).

main_info.md has been updated, as well as uispec, flag descriptions,
and flagspec. generator.py is where the binary patch is added.