Actually fixing -exp:moonbonus
In the last commit, I tried to load the effective "two byte map ID"
into the X register using "ldx $1701". That doesn't work, because
it puts the map ID ($1702) into the high byte instead of the low byte,
and the map plane ($1701) into the low byte instead of the high byte.
It basically gave you bonus experience everywhere except outdoors
on the overworld and underworld.

I've rewritten that part to actually put the correct values into X.