Softlock fix for -z:chaos wrt Black Hole
The f4c compiler does not appreciate 'wait' as the last command given
in a monster script; the partition_commands function in
compile_ai_script.py does something kind of weird if you try to put
an $FE byte at the end of the script and smushes the last two blocks
together, which breaks the script. Note that 'pass' is different from
'wait' (it's used as the specific do-nothing action instead of
signalling end-of-action/turn).

Unfortunately, in -z:chaos one possible situation is putting a Black
Hole at the end of the script, and as-written the f4c scripting says
"use Black Hole, show a message, wait", meaning the compiler would
output a broken script. The fix is to check if the attack script ends
with '    wait\n\n' and if so, simply remove that line.