Updating two treasure descriptions for spoilers
Two chests are somewhat incorrectly named for spoiler logs. They are:

1. The chest in Rydia's house that's past the first three you find.
In the vanilla game, it's a Change Rod; the vanilla FE spoiler log
called that the "Tiara room", when clearly it was not a Tiara.

It has been renamed to "second crawlspace room", since that's what
it is.

2. The chest in the part of Cave Eblan where you do a quick loop
between the west side of the first room after the town area and the
east side, that's through a secret path to the right. The map is
labelled "Pass to Bab-il (north connection)", and the chest in
question is listed as being in the "Exit" room despite not being able
to access it from the Rubi/Edge cutscene room without Push B to Jump.

It has been renamed to be in the north connection map, through a
secret path.