Adding more FuSoYa flags
This commit adds four new subflags for -fusoya, dealing with the way
FuSoYa un/learns spells and gains HP (as opposed to which spells he
learns and in what order). The randomizer-side logic is in
fusoya_rando.py, the flags logic is in flagsetcore.py and flagspec.txt,
the ui changes are in their usual spots, and the gameplay changes are
in fusoya_rando.py (for added f4c scripting), fusoya_challenge.f4c, and
character_expansion.f4c (for some unlearn changes). The flags are:

-fusoya:uncapped allows Fu to get up to 3900 HP; either he starts with
it, or he earns 100 HP from each boss in the game (possibly with the
cap being 3800 or 3700 under Bremove).

-fusoya:randomhp makes Fu gain HP from bosses randomly; sometimes he
will not gain any HP, and sometimes he will gain more than 100 HP at
one time. It is not fully random; for nerfed (1100 max) it can be at
most 200 HP per boss, for normal HP (1900 max) it can be at most 300,
and for uncapped (3700-3900 max) it can be up to 500. This flag is
implemented by using an HP gains array near the spell list; it turns
out that the vanilla FE code happily handles "give more than 100 HP"
for free because that's how Fu is initialized on-the-fly, so we just
pass an integer larger than 1 to get the numbers right.

-fusoya:slowstart makes Fu not gain spells and HP after 3 of the first
6 and 2 of the next 6 boss fights. The implementation uses the same
HP gains list from randomhp, only set up to guarantee missing HP gains
(via get_random_hp_gains in fusoya_rando.py). Without other HP flags,
under this flag Fu will hit full power after 19 bosses instead of 14.

-fusoya:unlearn is the "Benjamin Button" flag, where FuSoYa will start
with all of his spells and then *lose* spells after defeating bosses.
The order of spell loss is determined by the usual flag, and then
simply reversing the result; so, if the randomizer would make Fu learn
Nuke, Bersk, Cure4 after the 14th boss (let's assume j-spells), then
he will lose those spells after the *first* boss. Good luck? Under
-fusoya:location, the reversal does not happen, so Fu will lose spells
roughly correlated to the strength of the boss slot. The implementation
involves hooking into CharEx__GiveSpell in character_expansion.f4c and
when it's a Fu spellset pointing to FusoyaChallenge__TakeSpell_Impl
instead of the usual CharEx__GiveSpell_Impl.

The uispec for these flags is kind of obnoxious, because some subflags
only apply to some main categories of flag, so flagsetcore.py needed
some additions and the uispec was complicated.