Adding -fusoya:omnimage
This new flag gives FuSoYa a third spellset containing summons, Twin
magic, and Ninja magic, which he will learn the same way he learns
White and Black magic. The changes required for this to work are
somewhat non-trivial.

First, fusoya_rando.py has a number of changes in order to implement
the learning of new spells, as well as figuring out how many HP credits
are involved (which means Fu will gain more HP than normal, when
not using -fusoya:location).

Since this spellset is new, we need to ensure that the game recognizes
the new spellset as existing. We choose 0x0e for the spellset index,
which means we need to increase the number of spellsets that are
initialized by the game to include at least that many. It happens to
be the case that the maximum spellsets allowable as-is is 20 = 0x14,
so we remove the line in give_kain_magic.f4c that makes the game check
up to 0x0d (13) spellsets (having already changed MAX_SPELLSETS in
rom.py to 20). Then, in extend_spellsets.f4c, we set the total number
of spells checked to 24*20. (We also create a couple of labels for
Kain's spellsets, for ease of use.) generator.py is modified to add
this change is permanently (since it has no adverse effects; Kain's
0x01 spellset is unused normally and the game doesn't care, so we can
fill in a bunch of empty spellsets).

In fusoya_omnimage.f4c, we define FuSoYa's spellset, actually assign
that spellset to Fu for both menu and battle purposes, create "Omni"
text data for the menu to pull, write extra code to overwrite "Call"
in the menu (the same way "Ninja" overwrites "Black" for Edge), and
change Fu's battle menu when not under J-abilities. We choose the dummy
0x15 battle command (which is not used) for Omni, so we can just
overwrite the "Dummy" name in bank 0F with Omni and it works for free,
once we add in some code to check for Omni in battle the same way
the game checks for Ninja.

fusoya_challenge.f4c gets a small rewrite to (always) check for the
0x0e spellset, which does not matter when this flag is not on. There
is also a small change to include an inline substitution for initial HP
on -fusoya:vanilla, for each of handling credits.

character_expansion.f4c gets a modification to the -fusoya:unlearn
sections, to check for 0x0e as well as the other two spellsets.
japanese_abilities.f4c gets an inline substitution to replace Regen
with Omni (nothing lost there).

The flagspec, uispec, and flag descriptions have been updated.