Making Etoggle the default Encounters option
Evanilla being default is the bane of all our testers. That's been
changed now. To note, default flags are specified in the IMPLICIT
section of flagspec.txt, so this change flip-flops Evanilla with
Etoggle and reorders the two flags in the uispec, so that Etoggle
is at the top (unnecessary for default flags, but it looks nicer to
have the default option at the top). The dark grey box for Evanilla
stays, since that's the vanilla option colour.