Adding -tweak:cidairship
This flag adds a target-all Raid command for Cid, which does damage
based on his agility and the furthest airship you've obtained so far.
The bulk of the code is in cidairship.f4c, which adds the handlers
for the execution and audiovisuals for command $15 (formerly dummied
out, now useful!). There's also a helper function to check the plot
flags for the airships, and a hook in the timer calculation function
that sets a timer value like Jump does (in fact, it hooks in where
the Jump check happens).

The wacky f4cs that change command menus are changed to add Raid in
if necessary. Flagspec, uispec, flag descriptions, and the main_info.md
page have been updated. generator.py simply loads the relevant f4c
if necessary.