Adding -smith:omni
This flag changes the equip byte of the supersmith weapon, if it
exists, to allow every character that's in the seed to equip it. It
will not, however, allow characters to equip bows or arrows in case
the weapon is one of Rosa's. Oh well! At least Edge and Yang can equip
two-handed weapons.

Documentation/flagspec/uispec have been updated. The changes happen
in custom_weapon_rando.py, utilizing the same info as Tplayable.
flagsetcore.py is updated to prevent Chero from turning this flag off.