Fixing specs and adding Tadjmiabareas
Specs needed to be updated due to changes after the merge. So, that
happened.

Secondly, Scythe's fork of FE had a "vanilla miabs" adjustment to
the treasure weights for locations when -vanilla:miabs is on, to
help even the playing field for certain locations (e.g. Castle Eblan
is no longer strictly the best place to loot; Watery is similar!).
Galeswift's fork did not use that change. It has now been broken out
as a separate flag, Tadjmiabareas, which can be turned on independently
of -vanilla:miabs. The adjustment values are still the same.