Adding Kstart:zonk, updating to v4.6.2
In preparation for getting all of these flags to the public-facing site
in some way, updating the version to v4.6.2 (to ensure ease of
separation from the version the tournament will use).

Kstart:zonk removes the starting reward slot from the normal pool of
key item reward slots and places in the same tier as slots that can
only have non-key-item rewards. flagsetcore.py has been updated to
attempt to prevent bad scenarios from generating, because of a lack
of available key item reward slots. On certain combinations of flags,
nothing or a Cure1 will be placed in a reward slot instead of placing
e.g. the Spoon. (For example, on Tvanilla or Tshuffle.) Cross fingers
that this flag works correctly.