Adding Etable and -spoil:encounters
Thanks to leggy and ap7 for suggesting encounter randomization! This
family of flags randomizes the formations in the encounter groups
located at $0EC796-$0ECA95 (it does not randomize which groups belong
to which maps, partially because the Something Worth Fighting For
wacky already patches that area). There are five main categories of
randomization:
- shuffle[7/8] shuffles the formations within each group, possibly
leaving the Siren encounter alone (via shuffle7).
- relocated/ish shuffles the groups among each other, without changing
the order/rarity of the formations within the groups. So, you can
still Siren up a Yellow D or manip a D$ grind... once you find the
encounter group.
- wild/ish shuffle the formations outside of the groups they are found
in. wildish shuffles within four zones (overworld, underworld, moon,
and the Hook route/Giant), and wild combines the overworld/Hook/Giant
zones and the underworld/moon zones (since those groups all have the
same high-byte for encounter formations). Each formation shows up
as many times as it does in vanilla.
- chaos randomly assigns each encounter group eight formations chosen
independently (one at a time), so no formation is guaranteed to appear
and formations can be placed more often than usual. Formations with
high-byte 1 will appear in the underworld/moon groups, and high-byte 0
will appear in the overworld/Hook/Giant groups.
- uniquelygood assigns each encounter group one hand-picked formation
in all eight slots, which will be the same in every seed. These
formations are chosen because they are notable grinds or otherwise
useful or interesting encounters; no encounter appears in more than
one group.

All of the flags will report the encounters in the spoiler log, using
-spoil:encounters or -spoil:all (no obscuring possible via sparsity).

The randomization happens in encounter_rando.py, but the spoiler info
is in encounter_groups.csvdb and databases.py has been updated to
include this dbview and associated spoiler info function. All of the
relevant documention and specs have been updated.