Adding Kunreliabledark, fixing flag descriptions
Antidale made the default behaviour to implement "Klatedark" (or the
effect of that flag) 25% of the time, to make Darkness gated by
underground access 66% of the time, ish. That behaviour is now the
flag Kunreliabledark (named because sometimes Darkness is early, but
most of the time it's late).

Cspells:j,anti were missing ui flag descriptions, so those have been
updated too. (This commit follows the big merge of Anti's changes.)
The info page has also been updated.