Misc changes to wacky_rando, cidairship, etc.
Some of the wacky flags did not have "rom_address" as an input
to their apply function, so those have been updated.

Cid's Raid command had hooks at $03F210... which overlapped with
post_battle.f4c, on permadeath-type flags/etc. Those hooks have been
pushed back to $03F220. The rom map has been updated for this and
for the audiovisual hooks in bank02 for the same command.

There was a missing empty check in encounter_rando for the Etable flag.
That's been placed, so rolling seeds without Etable is now possible.