Tentative fix to gated objectives/doors rando
Issue 1: str(x) doesn't make sense for RewardSlot in Python 3.11 the
way that jayp had used it, so doors_rando has been rewritten a bit
to account for that and use the RewardSlots directly.

Issue 2: There's no "location" for the gated objective... so much like
the Crystal slot, we're skipping it in the validation. However! This
probably breaks logic somehow, so this commit is simply to give folks
a starting place to work from to account for this situation.