Building the boss_scaling_stats spreadsheet
boss_rando.py has had its "as main" function changed, to output
a .csv file of the boss scaling stats, together with instructions
for how to do that. The .csv is included in the main directory.

(In the event that this code makes it to a fork with e.g. Bstats:j,et,
simply tell the randomizer which formation data to load by changing
the "as main" function.)