Pass changes for Ogated
In vanilla FE, the Pass gets added to the key item pool (on Pkey)
right before the randomizer tries to fix a valid assignment, unlike
every actual key item, which get added much earlier. This fact
matters for Ogated, which removes the reward key item from the pool
after the key items are added... but not after the Pass is added.
This change adds the Pass in right after the key items are added,
so that it gets removed via Ogated if necessary.

Note that Kstart still behaves poorly re: Ogated. This change does
not impact those issues.

The only file changed is core_rando.py.