Fixing objective text for hard-req/gated
The '[key]' and '[crystal]' added to gated and hard-required
objectives, respectively, were not handled correctly for the purposes
of determining how long the lines were. So, textboxes in-game would
be off-centered (e.g. when you complete a hard-required objective) or
would overflow the tracker menu. This behaviour has been fixed, by
specifying specific string lengths for those lines in
objective_rando.py. There's a comment left above that to explain
what might happen in the future (maybe use ff4strlen instead of
hard-coding, or use a regex or something to fish out the relevant
lengths).

The "Gated objective now unlocked!" text was left-justified, and it
felt better for it to be centered, so that's done (in objectives.f4c).

Also, if it matters, there was a force-recompile, so hopefully
(cross fingers) all constants should work correctly now.