Tweaking Break Spell objective text
The "Break the Dark Elf's spell with the TwinHarp" objective text
is spaced in such a way that when it is hard-required or gated, the
extra key/crystal at the end overflows the tracker menu and looks bad.
So, this commit simply deletes the first "the" (before "Dark Elf"),
so that the objective now reads:

"Break Dark Elf's spell with the TwinHarp",

which does not have the same issue with the added characters.

The change happens in objectivespec.txt, which propagates to
objective_data.py (in two places) and uispec.js.