Fixing spoiler log for doors/entrances rando
Mist Cave and Mist Village have two entrances, going to different
sides of the map, and in doors_rando.py they have some tricksy
string manipulation going on in order to handle the spoiler log info.
... Unfortunately, "Mist" is a substring of "MistCave", so a case
was never being handled correctly (so Mist Cave had its entrances
overwritten with Mist East). That's been fixed, so that entrances
now lead to the correct place in the spoiler log. (The fix was just
to confirm that it's "Mist" and not "MistCave".)