Updating Shadow's spell list under Bwhichbez
The original set of spells for Shadow included "Disrupt"... but,
it was the invisible Disrupt that pre-Crystal Zeromus uses with the
shake to stop you from buffing your party before the actual fight.
That's been changed to be the instant-kill Disrupt used by TrapDoors
in the Sealed Cave... which means there's another reflectable
softlock opportunity. Watch out.

Demolish has also been added back in, since it fits the profile of
the spells that Shadow should be allowed to use.

The only changes are in golbez_rando.py.