Fixing up wacky_rando.py
In the Afflicted wacky, there is a table of 0x200 bytes that is
used for getting the status condition for a particular axtor and
formation. However, only the first half of the table is ever accessed,
so we replaced 0x200 with 0x100.

There was a try-except block left over from jayp's doors rando stuff
when it was still a wacky flag; that's been removed to just leave
the usual text.

The env.meta.get('wacky_challenge',[]) syntax sometimes had "None"
instead of "[]", so for consistency that's been changed everywhere.