Adding -tweak:twinmeteo
Wouldn't it be fun if Palom/Porom could cast W.Meteo? What if it
also came with a chance to self-target Stone, like in the vanilla
cutscene? This flag lets that happen (the self-target is just
the casting twins, not the whole party).

All changes are just in twin_meteo_stone.f4c.