Adding -wacky:skillissue
This wacky flag locks all commands except for Fight and Item, and
each boss defeated unlocks one of the commands. The changes are
fairly in-depth, given how many places check for the ability to
do certain commands.

The main functionality is in skillissue.f4c. We hook into the main
body of the code in three places (twice in the menu for drawing and
checking to cast spells, and once in battle code) and check if the
counter in wacky RAM is far enough for the given command. In the menus,
we also need to check for commands in order to cast spells; the fact
that this is really cumbersome under Misspelled made this a lot
harder but the result is fairly pretty. The menu checks are actually
in a wrapper because of how complex the logic ends up being, whereas
the battle check is a direct call.

In eventextensions_randomizer.f4c, a check is added before the FuSoYa
spell learning checks in Rando_PostBossBattle, where we increment
a counter in wacky RAM, determine which command is unlocked, and
display the textbox (which required writing a new LoadCommandName
function in text_buffers.f4c and taking up another text message,
which has been removed in unused.f4c).

In character_expansion.f4c, we add in a check only under this wacky
where right before pulling spellsets, we take note of the character's
job to allow Edge and Fu to have their special commands (Ninja/Omni)
to be distinguished from Black/Call.

In wacky_rando.py, we add in the apply_skillissue function, of course,
as well as the text data and whatnot, and note that this wacky takes
2 bytes of RAM. Just because, we remove a bunch of places where files
were being added again (mostly because Scythe didn't realize that they
were being added via the overall apply function).

Flagspec/etc. has been updated.