Tweaking the Egg kit
There was a "Sorc Robe" Egg kit, that was roughly "with Sorc Robe
and Cursed Ring and Glass Hat, you're probably invincible", but I
forgot that without Dragon-resist you are most certainly not
invincible against Yellow Dragons. So... that's been removed.
The MoonVeil version of the kit was given a BlackBelt, because
Rydia and Edward can be given weak weapons and with only one
attack multiplier might run out of Cure2s before the end of the fight.
(DKC will get a Black Sword, so it doesn't matter.)