Fixing an incorrect Skill Issue message
I had originally picked message $133 for the Skill Issue "unlocked
command" text. That's actually the Superhero textbox for being
weakened by gaining the Earth Crystal (not that it was noted
anywhere obvious, so if you forget to look, then you don't see it).
Instead, now message $134 is for Skill Issue, and is also correctly
loaded when it should be (instead of still calling $133).

Both changes are just in eventextensions_randomizer.f4c.