Adding the Work Experience wacky / updating etc.
The Work Experience wacky gives you an average-party-level-based
modifier to your experience from each fight, and then also gives
each axtor a random multiplier from .5 to 1.5 (increments of .25)
to their experience. The only changes are to wacky_rando.py
and experience_acceleration.f4c.

unused.f4c is annotated a bit to note some message usages, for ease
of reference.

This commit also updates flagspec/uispec, because 1. we need to
add the wacky (and make sure Skill Issue is labelled as a dangerous
and fork flag), and 2. we are removing -starting:underground and
-starting:blackchocobo, because Doors Rando doesn't currently work
with them and that's the currently existing set of flags. There is
also an update to main_info.md, to explain the situation with those
flags (and add Work Experience, of course).