Fixing Doors + Push B to Jump
The most recent version of Doors Rando was incompatible with
Push B to Jump, because too many triggers are added (it's exactly
one too many). The change here is to remove the trigger that Push B
adds to the Lower Bab-il revisit map that you can use to jump to
the Lugae side of the tower while you're headed down the Hook route,
when Doors Rando is also in play (either doors or entrances). In
addition, the wall that you jump over has been changed so that you
cannot jump over it, preventing access to a trigger tile with no
associated data (which would give you an underground inventory warp).