Letting the Rat Tail be auto-turned-in
The way the game checks if you hand in the Rat Tail is by checking
if it's the "selected item". But, if you happen to get the Pink Tail...
then it's now the "selected item", so that code immediately runs.
The change here is to copy/paste the Rat Tail code *again* to let
it run after the Pink Tail turn-in, if you happen to get it on Kpink.