Very small changes to Doors Rando
There was a situation where MistInn/etc. would be labelled wrong
because the comparison for entrances just asked if "Mist" is in the
string but not "MistCave", which hits the Inn and whatnot as well.
That's been changed. In addition, there was a missing comma in a
check for Damcyan being locked behind the omnicraft; that's also
been added.

The big print statement for the assignment data has also been removed,
not that this impacts people generating doors online.