Fixing softlocks/etc. with Skill Issue
It turns out that when Hide changes to Show, the flag for whether or
not the command should be disabled is not changed (because normally
Hide is never disabled), because Edward could auto-Hide while Hide
was still disabled, the game softlocks when pulling up his menu.

So, the solution chosen is to prevent auto-Hide when Hide is disabled;
makes sense! Similarly, we can prevent auto-Cover when Cover is
disabled; both of these are done in character_expansion.f4c. The Hide
change is done through ensuring that PreventAutoHideFlags is always
set for all characters and doesn't get cleared. The Cover change is
done through always setting CoverCapable to 0 for each character. The
Hide->Show thing is addressed in skillissue.f4c, and additional
Hide coding is in hide_softlock_fixes.f4c.

Finally, the table of "number of bosses defeated" in the ROM that
Skill Issue checks for each command is surrounded by FF, not 00, so
additional 00s are added at the end to ensure that Off, Show, etc. are
correctly enabled (because those can only be accessed when the
corresponding "first part" is enabled).