Fixing typo in Kit Rando
The Dragoon Armor's FE-internal name is #item.DragonArmor, but
in kit_rando.py an Egg kit option had it "DragoonArmor", which would
throw a TypeError during generation. That typo has been fixed.

(Should it be "DragonArmor"? Eh, not for me to decide. Some of the
items seem to have had names taken from different versions.)