Updating flagspec/etc. and flagsetcore.py
This commit fixes a few flagspec bugs... or, tries to (pending
the merging of some string formatting issues in flagsetcore regarding
rf/raw-formatted strings and the javascripthon transpiler).

1. Skill Issue had not been included in the mutual exclusion with
Darts and Musical. That's been fixed.

2. Tstandardish had not been considered a weighted distribution for
the purposes of Tadjmiabareas in flagsetcore; that's been added.

3. Bremove:[...] has been broken the whole time in the ui, because it
had been typoed there (and in uispec_flagdescriptions). That's been
fixed.

4. In flagsetcore, there was code to count objectives in order to
ensure that you were not trying to fight more bosses than there are
bosses in the seed, and similar. This code didn't work, because of
weird javascripthon transpiler issues. Thanks to Wylem and Galeswift
for fixing that stuff... eventually, since this commit features a
couple of changes to flagsetcore that should handle Bremove correctly,
and the other changes will come later. To confirm.