Fix Ctreasure miab crash
A value pushed to the stack was not being pulled before rtl'ing after a successful "it's a character reward" check from a miab, in rewards.f4c. That's been changed (and the zonk counter is correctly incremented if it should be, which was the point of pushing that value to the stack earlier).