Fixing Kmiab Pnone Tvanilla/etc generation issue
On Tshuffle and Tvanilla with Pnone, Baron Castle does not give a
key item reward, which means there is supposed to be an empty reward
that gets shuffled into the pool. However, treasure_rando.py tries
to write spoiler log data for an empty reward in a miab, which fails.

While we could write special cases to put a Cure1 there, it is mostly
easier to just remove the empty reward entirely and always just
place a Cure1 where normally there would be nothing (which is the
same behaviour, roughly, as when Tplayable cannot place items in
miabs). So, that's the change in core_rando.py.