Adding guaranteed apples to Kokkol when needed
On Sstandard and Spro, Salways:apples would not actually put apples
anywhere, because those items can only go into Kokkol's shop and
that shop is handled before the rest of the guaranteed items. So,
this change to shop_rando.py adds in a random choice of either the
AuApple or the AgApple (bypassing Sno:j, as Salways normally does).