Another fix for characters in miabs
The $c6 value was not being cleared when getting a character from a
miab on Ctreasure, so the game tried to keep giving you an item
(usually a Cure1, sometimes nothing for some reason), and the battle
background seems to not clear correctly. Both of these issues get
fixed when you simply clear $c6 in rewards.f4c.