Modifying extract.py to get more equips info
extract.py has been modified to output lists of items for each
entry in the "which jobs can equip these items?" table, for the purpose
of easily identifying which items use which entries. The results are
in mod_extracted_info.txt. The main purpose for this change was to
more easily implement -tweak:rosadin.